Tweakes types & jest config
