+++ packages/contentful-ssg/package.json
@@ -56,8 +56,8 @@
     }
   },
   "bin": {
-    "cssg": "./dist/index.js",
-    "contentful-ssg": "./dist/index.js"
+    "cssg": "./dist/cli.js",
+    "contentful-ssg": "./dist/cli.js"
   },
   "scripts": {
     "clean": "rimraf ./dist",
@@ -136,7 +136,6 @@
     }
   },
   "eslintConfig": {
-    "root": true,
     "extends": [
       "xo",
       "xo-typescript/space"
@@ -144,7 +143,7 @@
     "parser": "@typescript-eslint/parser",
     "parserOptions": {
       "project": [
-        "./tsconfig.json"
+        "tsconfig.json"
       ]
     },
     "ignorePatterns": [
