+++ packages/contentful-ssg/src/helper/create-require.ts
@@ -1,0 +1,3 @@
+import {createRequire as moduleCreateRequire} from 'module';
+
+export const createRequire = (rootDir:string = null) => rootDir === null ? moduleCreateRequire(import.meta.url) : moduleCreateRequire(`${rootDir}/:internal:`);
