+++ packages/contentful-ssg/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "outDir": "dist"
   },
-  "include": ["src/**/*.ts"],
+  "include": ["src/index.ts","src/**/*.ts"],
   "exclude": ["src/**/*.test.ts"]
 }
